Address review feedback and read atomic FileSegment fields locklessly
Review fixes for #108477:
- `shrinkFileSegmentToDownloadedSize`: when a partial segment becomes
  `DOWNLOADED`, free the whole `download_data`, not just the nested
  reader/writer. Otherwise the per-download allocation leaks on an
  already-cached segment and the `!download_data` invariant (asserted in
  `assertCorrectnessUnlocked`) is violated.
- Keep `hits_count`/`ref_count` at 64-bit: `hits_count` backs the `UInt64`
  `system.filesystem_cache.cache_hits` column and must not silently wrap on
  long-running servers.

Contention: read already-atomic fields (`state`, `getReservedSize`,
`isDownloaded`, `isDetached`) without taking the per-segment lock - each
returns a single-field snapshot, so the lock added no consistency and only
contended hot segments.

+++ src/Interpreters/FileCache/FileSegment.h
@@ -315,8 +315,8 @@ private:
     LoggerPtr log;
 #endif
 
-    std::atomic<UInt32> hits_count = 0; /// cache hits.
-    std::atomic<UInt32> ref_count = 0; /// Used for getting snapshot state
+    std::atomic<size_t> hits_count = 0; /// cache hits.
+    std::atomic<size_t> ref_count = 0; /// Used for getting snapshot state
 
     /// Guarded by `segment_guard`. Set while dynamic-resize eviction is pending.
     bool on_delayed_removal = false;
